+++ tests/Phinx/Console/Command/CreateTest.php
@@ -58,8 +58,10 @@ class CreateTest extends \PHPUnit_Framework_TestCase
             )
         );
 
-        foreach (glob($this->config->getMigrationPath().'/*.*') as $migration) {
-            unlink($migration);
+        foreach ($this->config->getMigrationPaths() as $path) {
+            foreach (glob($path.'/*.*') as $migration) {
+                unlink($migration);
+            }
         }
 
         $this->input = new ArrayInput([]);
@@ -79,7 +81,9 @@ class CreateTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('create');
 
         /** @var Manager $managerStub mock the manager class */
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
+        $managerStub = $this->getMockBuilder('\Phinx\Migration\Manager')
+            ->setConstructorArgs(array($this->config, $this->input, $this->output))
+            ->getMock();
 
         $command->setConfig($this->config);
         $command->setManager($managerStub);
@@ -104,7 +108,9 @@ class CreateTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('create');
 
         /** @var Manager $managerStub mock the manager class */
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
+        $managerStub = $this->getMockBuilder('\Phinx\Migration\Manager')
+            ->setConstructorArgs(array($this->config, $this->input, $this->output))
+            ->getMock();
 
         $command->setConfig($this->config);
         $command->setManager($managerStub);
@@ -126,7 +132,9 @@ class CreateTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('create');
 
         /** @var Manager $managerStub mock the manager class */
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
+        $managerStub = $this->getMockBuilder('\Phinx\Migration\Manager')
+            ->setConstructorArgs(array($this->config, $this->input, $this->output))
+            ->getMock();
 
         $this->config['templates'] = array(
             'file' => 'MyTemplate',
@@ -186,7 +194,9 @@ class CreateTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('create');
 
         /** @var Manager $managerStub mock the manager class */
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
+        $managerStub = $this->getMockBuilder('\Phinx\Migration\Manager')
+            ->setConstructorArgs(array($this->config, $this->input, $this->output))
+            ->getMock();
 
         foreach ($config as $key => $value) {
             $this->config[$key] = $value;
@@ -248,7 +258,9 @@ class CreateTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('create');
 
         /** @var Manager $managerStub mock the manager class */
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
+        $managerStub = $this->getMockBuilder('\Phinx\Migration\Manager')
+            ->setConstructorArgs(array($this->config, $this->input, $this->output))
+            ->getMock();
 
         foreach ($config as $key => $value) {
             $this->config[$key] = $value;
@@ -309,7 +321,9 @@ class CreateTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('create');
 
         /** @var Manager $managerStub mock the manager class */
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
+        $managerStub = $this->getMockBuilder('\Phinx\Migration\Manager')
+            ->setConstructorArgs(array($this->config, $this->input, $this->output))
+            ->getMock();
 
         foreach ($config as $key => $value) {
             $this->config[$key] = $value;