+++ tests/Phinx/Util/UtilTest.php
@@ -85,4 +85,32 @@ class UtilTest extends \PHPUnit_Framework_TestCase
             $this->assertEquals($expectedResult, Util::isValidPhinxClassName($input));
         }
     }
+
+    public function testGlobPath()
+    {
+        $files = Util::glob(__DIR__ . '/_files/migrations/empty.txt');
+        $this->assertCount(1, $files);
+        $this->assertEquals('empty.txt', basename($files[0]));
+
+        $files = Util::glob(__DIR__ . '/_files/migrations/*.php');
+        $this->assertCount(3, $files);
+        $this->assertEquals('20120111235330_test_migration.php', basename($files[0]));
+        $this->assertEquals('20120116183504_test_migration_2.php', basename($files[1]));
+        $this->assertEquals('not_a_migration.php', basename($files[2]));
+    }
+
+    public function testGlobAll()
+    {
+        $files = Util::globAll(array(
+            __DIR__ . '/_files/migrations/*.php',
+            __DIR__ . '/_files/migrations/subdirectory/*.txt'
+        ));
+
+        $this->assertCount(4, $files);
+        $this->assertEquals('20120111235330_test_migration.php', basename($files[0]));
+        $this->assertEquals('20120116183504_test_migration_2.php', basename($files[1]));
+        $this->assertEquals('not_a_migration.php', basename($files[2]));
+        $this->assertEquals('empty.txt', basename($files[3]));
+
+    }
 }